requirements.txt
aiogram==2.24
matplotlib==3.4.3
numpy==1.23.3
Pillow==9.4.0
tensorflow==2.11.0
tensorflow_hub==0.12.0
tensorflow_intel==2.11.0
torch==1.12.1
torchvision==0.13.1
